Move RefactoringRelationShip to src root due to the dependencies in the source files Fix for https://github.com/tsantalis/RefactoringMiner/issues/382
