+++ src-test/org/refactoringminer/utils/RefactoringRelationshipGroup.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.refactoringminer.api.RefactoringType;
+import org.refactoringminer.util.RefactoringRelationship;
 
 public class RefactoringRelationshipGroup {
